todos los get all funcionan estan probados en postman

+++ Backend/Maen/src/main/java/com/triana/salesianos/dam/Maen/dto/supermarket/GetSupermarketDTO.java
@@ -10,14 +10,14 @@ public record GetSupermarketDTO(
         UUID id,
         String name,
         String logotype,
-        int numOfProduct
+        Long numOfProduct
 ) {
     public static GetSupermarketDTO of (SuperMarket supermarket){
         return GetSupermarketDTO.builder()
                 .id(supermarket.getId())
                 .name(supermarket.getName())
                 .logotype(supermarket.getLogotype())
-                .numOfProduct(0)
+                .numOfProduct(0L)
                 .build();
     }
 }
